+++ apps/drizzle-run/app/registry/dialects/postgresql/presets/rls/manifest.json
@@ -1,5 +1,5 @@
 {
   "$schema": "../../../manifest.schema.json",
-  "name": "Starter 01",
-  "description": "A starter demo"
+  "name": "(Experimental) Row Level Security (RLS)",
+  "description": "A demo of RLS with a Supabase emulator"
 }
