+++ apps/drizzle-run/app/registry/dialects/postgresql/presets/starter-01/index.ts
@@ -3,54 +3,15 @@
  * 💡Tip: you can use the `$` global variable to access goodies
  */
 
-import { eq } from "drizzle-orm";
-import { drizzle } from "drizzle-orm/pglite";
+import { db } from "./db";
+import {} from "./schema";
 
-import { db as admin } from "./db";
-import * as schema from "./schema";
-import { createDrizzle } from "./utils";
+// 💡 This playground run a seed code in the Seed tab
 
-const userId1 = $.random.uuid();
-const userId2 = $.random.uuid();
-
-const client = drizzle({
-  client: admin.$client,
-  schema,
-  casing: "snake_case",
+const result = await db.query.users.findMany({
+  with: {
+    posts: true,
+  },
 });
 
-const db = createDrizzle({ role: "authenticated", sub: userId1 }, { admin, client });
-
-await db.admin.insert(schema.authUsers).values([
-  {
-    id: userId1,
-  },
-  {
-    id: userId2,
-  },
-]);
-
-await db.admin.insert(schema.profiles).values({
-  id: userId1,
-  email: "demo@mail.com",
-});
-
-console.log(
-  "insert allowed",
-  await db.rls(async (tx) => tx.insert(schema.rooms).values({ topic: "user room" }).returning()),
-);
-
-console.log(
-  "user can not edit other profile. Before",
-  await db.admin.query.profiles.findMany(),
-  await db.rls(async (tx) => {
-    await tx.update(schema.profiles).set({ email: "nope" }).where(eq(schema.profiles.id, userId2));
-  }),
-  "after",
-  await db.admin.query.profiles.findMany(),
-);
-
-const dbAnonymous = createDrizzle({ role: "anon" }, { admin, client });
-
-console.log("anonymous can not read");
-console.log(await dbAnonymous.rls(async (tx) => tx.select().from(schema.rooms)));
+console.log(result);
