+++ apps/drizzle-run/app/registry/dialects/postgresql/presets/starter-01/manifest.json
@@ -1,5 +1,5 @@
 {
   "$schema": "../../../manifest.schema.json",
-  "name": "(Experimental) Row Level Security (RLS)",
-  "description": "A demo of RLS with a Supabase emulator"
+  "name": "Simple example",
+  "description": "A simple example to start with"
 }
