+++ apps/drizzle-run/app/registry/dialects/sqlite/presets/starter-01/manifest.json
@@ -1,5 +1,5 @@
 {
   "$schema": "../../../manifest.schema.json",
-  "name": "Starter 01",
-  "description": "A starter demo"
+  "name": "Simple example",
+  "description": "A simple example to start with"
 }
