+++ apps/drizzle-run/package.json
@@ -41,8 +41,8 @@
     "@electric-sql/pglite": "^0.2.3",
     "@electric-sql/pglite-v1": "npm:@electric-sql/pglite@0.1.5",
     "@epic-web/cachified": "^5.2.0",
-    "@libsql/client-wasm": "^0.14.0",
     "@libsql/client": "^0.14.0",
+    "@libsql/client-wasm": "^0.14.0",
     "@monaco-editor/react": "^4.6.0",
     "@paralleldrive/cuid2": "^2.2.2",
     "@remix-run/node": "^2.14.0",
@@ -55,7 +55,8 @@
     "@xstate/react": "^4.1.1",
     "@xstate/store": "^2.6.0",
     "drizzle-kit": "^0.28.1",
-    "drizzle-orm": "^0.36.3",
+    "drizzle-orm": "^0.36.4",
+    "drizzle-seed": "^0.1.1",
     "drizzle-valibot": "^0.2.0",
     "drizzle-zod": "^0.5.1",
     "isbot": "^4.1.0",
